Time: 1 ms (13.52%) | Memory: 9.6 MB (77.86%) - LeetSync

+++ 2812-find-the-maximum-achievable-number/find-the-maximum-achievable-number.cpp
@@ -1,6 +1,6 @@
 class Solution {
 public:
-    short theMaximumAchievableX(short num, short t) {
+    long long theMaximumAchievableX(long long num, long long t) {
         return num + t + t;
     }
 };
